pyproject.toml
[project]
name = "spite-ai"
version = "0.1.0"
description = "AI system that responds as a user from Spite Magazine using semantic search and Mistral"
readme = "README_IMPROVEMENTS.md"
requires-python = ">=3.11"
authors = [
    {name = "Jaden", email = "jaden@example.com"}
]
keywords = ["ai", "nlp", "semantic-search", "mistral", "spite-magazine"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.3",
    "sentence-transformers>=5.1.0",
    "ollama>=0.3.0",
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "groq>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
spite-ai = "spite_ai:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spite_ai"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "W503", "S101", "PLR0913", "PLR0912", "PLR0915"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
